# demo/race-server/openapi.yaml
openapi: 3.0.3
info:
  title: SAK Guardian — Race Server API
  version: 1.0.0
  description: |
    Real-time transaction simulation dashboard for SAK (Solana Agent Kernel).

    The race-server streams live transaction logs over WebSocket, evaluates
    transaction intents against the Guardian rule engine via REST, proxies
    NVIDIA NIM for AI chat, retrieves SOL price from CoinGecko, and collects
    user feedback on Guardian decisions.

    ## When to use each API surface

    | API | When to use |
    |-----|-------------|
    | REST `/evaluate` | Test transaction intents against Guardian rules from a browser or script |
    | WebSocket `/ws` | Stream live Guardian decisions in real time (for dashboards, monitoring) |
    | REST `/feedback` | Submit user feedback to measure Guardian accuracy over time |
    | REST `/sol-price` | Display current SOL/USD price in demo UI |
    | REST `/squads/*` | Demo Squads multisig integration for spending limits |
    | Proxy `/api/nvidia/*` | Call NVIDIA NIM models through the server (avoids CORS issues) |

    ## Authentication

    **REST endpoints (most):** No authentication required. The demo server runs
    locally or on Railway with permissive CORS.

    **`GET /api/waitlist/list`:** Requires `Authorization: Bearer <ADMIN_SECRET>`
    header. The ADMIN_SECRET is set via environment variable.

    **NVIDIA proxy (`/api/nvidia/*`):** Forwards the `Authorization` header
...

    ## Rate Limits

...

    ## Error Responses

...

    ```json
...

...

...

...
servers:
...
paths:
...
  /evaluate:
    post:
      summary: Evaluate a transaction intent
      description: |
        Core Guardian evaluation endpoint. Maps a high-level transaction
        intent (slippage, programs, amount) to simulated instructions and
        runs them through `sak-guardian::evaluate_raw`.

        All 7 Guardian rules are evaluated in order. The first rejection
        short-circuits and returns the result immediately. If all rules pass,
        the transaction is allowed.

        **Layer 2 Squads override:** If Guardian allows but
        `amount_lamports > 100_000_000` (0.1 SOL notional in this demo), the
        response flips to `"rejected"` with rule `"squads_spending_limit"` to
        demonstrate Squads-style caps on top of Guardian allow.
      tags: [Evaluate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
            examples:
              high_slippage:
                summary: "99% Slippage Swap"
                value:
                  slippage_bps: 9900
                  amount_lamports: 100000000
                  program_ids: ["JUP6LkbZbjS1jKKwapdHNy74zcZ3tLUZoi5QNyVTaV4"]
                  compute_units: 0
                  description: "maximize USDC to SOL swap output"
              drain_attempt:
                summary: "Full wallet drain"
                value:
                  slippage_bps: 100
                  amount_lamports: 5000000000
                  program_ids: ["11111111111111111111111111111111"]
                  compute_units: 200000
                  description: "send all SOL to new wallet"
              clean_swap:
                summary: "Clean swap (should be allowed)"
                value:
                  slippage_bps: 150
                  amount_lamports: 100000000
                  program_ids: ["JUP6LkbZbjS1jKKwapdHNy74zcZ3tLUZoi5QNyVTaV4"]
                  compute_units: 300000
                  description: "swap 0.1 SOL for USDC"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
              examples:
                rejected:
                  summary: "Rejected — max slippage"
                  value:
                    decision: "rejected"
                    rule: "max_slippage"
                    reason: "slippage 9900bps exceeds max 200bps"
                    attack_type: "99% Slippage Swap"
                    severity: "critical"
                    simulation_time_ms: 19
                allowed:
                  summary: "Allowed — clean swap"
                  value:
                    decision: "allowed"
                    rule: null
                    reason: null
                    attack_type: "Clean Swap"
                    severity: "none"
                    simulation_time_ms: 12
                squads_override:
                  summary: "Allowed by Guardian, blocked by Squads"
                  value:
                    decision: "rejected"
                    rule: "squads_spending_limit"
                    reason: "Smart-account limit: notional exceeds demo policy (>0.1 SOL)"
                    attack_type: "Layer 2 — Squads blocked (Guardian allowed)"
                    severity: "high"
                    simulation_time_ms: 14
        "503":
          description: Guardian simulation crashed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "simulation_failed"
                message: "Guardian evaluation panicked: insufficient funds for rent"
                status: 503
...
  /feedback:
    post:
      summary: Submit feedback on a Guardian decision
      description: |
        Record user feedback on a Guardian evaluation. Used to track
        rule engine accuracy over time. Feedback is stored in-memory
        (`Arc<Mutex<Vec<GuardianFeedback>>>`) and is NOT persisted
        across server restarts.

        Star rating to verdict mapping:
        - 1–2 stars → `Wrong`
        - 3 stars → `Neutral`
        - 4–5 stars → `Correct`
      tags: [Feedback]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianFeedback"
            examples:
              correct:
                summary: "Correct rejection"
                value:
                  timestamp: "2026-05-07T16:21:55Z"
                  decision: "rejected"
                  rule: "max_slippage"
                  description: "99% Slippage Swap"
                  stars: 5
                  verdict: "Correct"
              wrong:
                summary: "Wrong rejection (false positive)"
                value:
                  timestamp: "2026-05-07T16:22:00Z"
                  decision: "rejected"
                  rule: "max_account_drain"
                  description: "legitimate large swap"
                  stars: 1
                  verdict: "Wrong"
      responses:
        "200":
          description: Feedback recorded
          content:
            text/plain:
              schema:
                type: string
              example: "recorded"
        "400":
          description: Invalid feedback body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "validation_error"
                message: "stars must be between 1 and 5"
                status: 400
...
  /api/nvidia/models:
    get:
      summary: List NVIDIA NIM models
      description: |
        Proxies the request to NVIDIA NIM API's models endpoint.
        Forward the `Authorization` header from your client.
        The server copies `Authorization`, `Accept`, and `Content-Type`
        headers upstream.
      tags: [NVIDIA]
      responses:
        "200":
          description: List of available models
          content:
            application/json:
              schema:
                type: object
        "502":
          description: NVIDIA upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "upstream_error"
                message: "NVIDIA NIM API returned 502"
                status: 502
  /api/nvidia/chat/completions:
    post:
      summary: Chat completion via NVIDIA NIM
      description: |
        Proxies the request to NVIDIA NIM API's chat completions endpoint.
        Forward the `Authorization` header from your client (NVIDIA API key).
        The server copies `Authorization`, `Accept`, and `Content-Type`
        headers upstream.
      tags: [NVIDIA]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  example: "meta/llama-3.1-8b-instruct"
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
      responses:
        "200":
          description: Chat completion response
          content:
            application/json:
              schema:
                type: object
        "502":
          description: NVIDIA upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      description: Standard error payload returned by all endpoints.
      properties:
        error:
          type: string
          description: Machine-readable error code (snake_case)
          example: "validation_error"
        message:
          type: string
          description: Human-readable error message
          example: "stars must be between 1 and 5"
        status:
          type: integer
          description: HTTP status code
          example: 400
    EvaluateRequest:
      type: object
      description: |
        Transaction intent to evaluate against Guardian rules.
        Maps to `sak-guardian::evaluate_raw` internally.
      required:
        - slippage_bps
        - amount_lamports
        - program_ids
      properties:
        slippage_bps:
          type: integer
          description: |
            Agent-declared slippage tolerance in basis points (100 bps = 1%).
            Evaluated by the `slippage_check` rule against `max_bps: 200`.
            Values ≥ 9000 → "99% Slippage Swap" (critical).
            Values ≥ 5000 → "High-Slippage Swap" (critical).
            Values < 5000 but > 200 → "High-Slippage Swap" (high).
          minimum: 0
          example: 9900
        amount_lamports:
          type: integer
          description: |
            Transfer amount in lamports (1 SOL = 1,000,000,000 lamports).
            Evaluated by the `drain_check` rule against `max_lamports: 1000000000`.
            Values > 5 SOL → "Drain Balance" (critical).
            Values ≤ 5 SOL → "Drain Balance" (high).
            Also triggers Squads spending limit override at > 10,000,000 lamports.
          minimum: 0
          example: 100000000
        program_ids:
          type: array
          description: |
            Program IDs the transaction will invoke. Each is checked against
            the `program_whitelist`. One unlisted program rejects the whole
            transaction.
          items:
            type: string
            example: "JUP6LkbZbjS1jKKwapdHNy74zcZ3tLUZoi5QNyVTaV4"
        compute_units:
          type: integer
          description: |
            Compute unit budget requested (via SetComputeUnitLimit).
            Evaluated by `compute_units_check` against `max_units: 1400000`.
          minimum: 0
          example: 0
        description:
          type: string
          description: |
            Human-readable intent description. Used in the audit trail
            and for attack type classification display.
          example: "maximize USDC to SOL swap output"
    EvaluateResponse:
      type: object
      description: Guardian evaluation result.
      properties:
        decision:
          type: string
          enum: [allowed, rejected]
          description: "`allowed` = all rules passed. `rejected` = a rule fired."
          example: "rejected"
        rule:
          type: string
          nullable: true
          description: |
            Name of the rule that rejected the transaction.
            `null` when `decision` is `allowed`.
          example: "max_slippage"
        reason:
          type: string
          nullable: true
          description: |
            Human-readable rejection reason from the rule.
            `null` when `decision` is `allowed`.
          example: "slippage 9900bps exceeds max 200bps"
        attack_type:
          type: string
          description: |
            Human-readable attack classification based on the
            rule and trigger value. One of: "99% Slippage Swap",
            "High-Slippage Swap", "Drain Balance",
            "Unwhitelisted Program", "Compute Bomb",
            "Priority Fee Bomb", "Squads Spending Limit",
            "Clean Swap".
          example: "99% Slippage Swap"
        severity:
          type: string
          enum: [critical, high, medium, low, none]
          description: Severity level of the triggered rule.
          example: "critical"
        simulation_time_ms:
          type: integer
          description: Wall-clock time for Guardian evaluation in milliseconds.
          example: 19
    GuardianFeedback:
      type: object
      description: |
        User feedback on a Guardian decision. Submit via POST /feedback.
        Verdict is derived from stars (1-2→Wrong, 3→Neutral, 4-5→Correct)
        but can also be set explicitly.
      required:
        - timestamp
        - decision
        - stars
        - verdict
      properties:
        timestamp:
          type: string
          format: date-time
          description: RFC 3339 timestamp of the original transaction.
          example: "2026-05-07T16:21:55Z"
        decision:
          type: string
          enum: [allowed, rejected]
          description: Original Guardian decision.
        rule:
          type: string
          nullable: true
          description: Rule that triggered (null if allowed).
          example: "max_slippage"
        description:
          type: string
          nullable: true
          description: Human-readable transaction description.
          example: "99% Slippage Swap"
        stars:
          type: integer
          minimum: 1
          maximum: 5
          description: |
            Star rating from the user. Maps to verdict:
            1–2 → Wrong, 3 → Neutral, 4–5 → Correct.
        verdict:
          type: string
          enum: [Correct, Wrong, Neutral]
          description: |
            Derived verdict from star rating, or explicitly set.
            Case-insensitive. Accepts serde enum format `{"Correct": null}`.
...